fixed redirections
