+++ srcs/get_input/parsing.c
@@ -6,7 +6,7 @@
 /*   By: hzimmerm <hzimmerm@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/17 11:49:53 by Henriette         #+#    #+#             */
-/*   Updated: 2024/09/09 20:03:50 by hzimmerm         ###   ########.fr       */
+/*   Updated: 2024/09/09 21:16:43 by hzimmerm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,6 @@ int	parse_line(char *cmd_line, t_input **command, t_global *global)
 	}
 	free_array(elmts.array);
 	expand_var_words(*command, global);
-	//print_arrays_testing(command);
 	return (0);
 }
